+++ doc/changes.rst
@@ -16,7 +16,13 @@ desitarget Change Log
         * Routines to read and match to the lightweight files also added.
         * Can be used to flag secondaries that are close to bright stars.
 
+* Circumvent warnings when running with Python >3.12 [`PR #870`_].
+    * Should address `issue #865`_ and `issue #866`_.
+
 .. _`issue #503`: https://github.com/desihub/fiberassign/issues/503
+.. _`issue #865`: https://github.com/desihub/desitarget/issues/865
+.. _`issue #866`: https://github.com/desihub/desitarget/issues/866
+.. _`PR #870`: https://github.com/desihub/desitarget/pull/870
 .. _`PR #877`: https://github.com/desihub/desitarget/pull/877
 
 4.6.0 (2026-01-30)